Fix some build warnings (#5941)
* Fix some build warnings

I added the -Wno-c99-designator flag because we were getting a warning
about syntax that requires c99 compatibility (which I guess isn't
exactly a subset of C++20). However, as far as I can tell, clang and gcc
support the syntax, so the warning isn't very useful to us at this point
in time.

* [run CI]

+++ bodo/io/iceberg_parquet_write.cpp
@@ -410,7 +410,7 @@ void iceberg_pq_write(const char *table_data_loc,
                 // See function description string for more details.
                 p.iceberg_file_info_py = PyTuple_New(3 + num_keys);
 
-                for (size_t j = 0; j < num_keys; j++) {
+                for (int64_t j = 0; j < num_keys; j++) {
                     auto transformed_part_col = transform_cols[j];
                     // convert transformed partition value to string
                     std::string value_str = transform_val_to_str(
